# PERF: DataFrame.astype where dtype is an ExtensionDtype