--- a/asv_bench/benchmarks/frame_methods.py
+++ b/asv_bench/benchmarks/frame_methods.py
@@ -17,6 +17,40 @@ from pandas import (
 from .pandas_vb_common import tm
 
 
+class AsType:
+    params = [
+        [
+            # from_dtype == to_dtype
+            ("Float64", "Float64"),
+            ("float64[pyarrow]", "float64[pyarrow]"),
+            # from non-EA to EA
+            ("float64", "Float64"),
+            ("float64", "float64[pyarrow]"),
+            # from EA to non-EA
+            ("Float64", "float64"),
+            ("float64[pyarrow]", "float64"),
+            # from EA to EA
+            ("Int64", "Float64"),
+            ("int64[pyarrow]", "float64[pyarrow]"),
+        ],
+        [False, True],
+    ]
+    param_names = ["from_to_dtypes", "copy"]
+
+    def setup(self, from_to_dtypes, copy):
+        from_dtype = from_to_dtypes[0]
+        if from_dtype in ("float64", "Float64", "float64[pyarrow]"):
+            data = np.random.randn(100, 100)
+        elif from_dtype in ("int64", "Int64", "int64[pyarrow]"):
+            data = np.random.randint(0, 1000, (100, 100))
+        else:
+            raise NotImplementedError
+        self.df = DataFrame(data, dtype=from_dtype)
+
+    def time_astype(self, from_to_dtypes, copy):
+        self.df.astype(from_to_dtypes[1], copy=copy)
+
+
 class Clip:
     params = [
         ["float64", "Float64", "float64[pyarrow]"],